refactor: decouple request client from rabbitmq
